Added solution - LeetHub

+++ Sum of Series - GFG/sum-of-series.cpp
@@ -10,7 +10,7 @@ public:
 	// function to return sum of  1, 2, ... n
 	long long seriesSum(int n) {
 	    long long int sum;
-	    sum = (n * 1ll * (n + 1))/2;
+	    sum = (long long)((n * 1ll * (n+1))/2);
 	    return sum;
 	}
 };
